fix: CreatePayment의 customerId Integer로 수정

+++ src/main/java/staysplit/hotel_reservation/payment/domain/dto/request/CreatePaymentRequest.java
@@ -3,7 +3,7 @@ package staysplit.hotel_reservation.payment.domain.dto.request;
 import staysplit.hotel_reservation.payment.domain.enums.PaymentStatus;
 
 public record CreatePaymentRequest(
-        Long customerId,
+        Integer customerId,
         String payMethod,
         String payName,
         Integer amount,
